Update divide MarkerListComponent by MapComponent
